// src/types/market.ts
export interface MarketData {
  price: number;
  priceChange: number;
  priceChangePercent: number;
  highPrice: number;
  lowPrice: number;
  volume: number;
  quoteVolume: number;
  marketCap: number;
  dominance: number;
  circulatingSupply: number;
  totalSupply: number;
  maxSupply: number;
  lastUpdate: number;
} 